Move `IntervalIdentifier` into separate `interfaces` module.
